requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain-anthropic
python-dotenv
cx_Freeze
ttkbootstrap==0.5.1
pillow
playwright
chardet
requests
